+++ CHANGES.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
++ core: Don't require `Clone` and `Debug` for the generic action name parameter in `RelmAction`
+
 ## 0.8.1 - 2024-3-13
 
 ### Fixed